chatbot/requirements.txt
langchain
langchain-openai
black
python-dotenv
fastapi
uvicorn
psycopg2